Restyled command modal.
