+++ packages/storybook/decorators/SubQueryDecorator.tsx
@@ -1,0 +1,9 @@
+import { DecoratorFn } from '@storybook/react'
+
+import { SubQueryProvider } from '@dao-dao/state'
+
+export const SubQueryDecorator: DecoratorFn = (Story) => (
+  <SubQueryProvider>
+    <Story />
+  </SubQueryProvider>
+)
