+++ packages/storybook/decorators/index.tsx
@@ -6,5 +6,6 @@ export * from './makeProposalModuleAdapterDecorator'
 export * from './NotificationsDecorator'
 export * from './ReactHookFormDecorator'
 export * from './RecoilDecorator'
+export * from './SubQueryDecorator'
 export * from './ThemeDecorator'
 export * from './WalletProviderDecorator'
