+++ packages/ui/components/SearchBar.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentMeta, ComponentStory } from '@storybook/react'
 import { SearchBar } from './SearchBar'
 
 export default {
-  title: '(OLD DAO DAO) / components / SearchBar',
+  title: 'DAO DAO / packages / ui / components / SearchBar',
   component: SearchBar,
 } as ComponentMeta<typeof SearchBar>
 
@@ -13,3 +13,8 @@ const Template: ComponentStory<typeof SearchBar> = (args) => (
 
 export const Default = Template.bind({})
 Default.args = {}
+
+export const Ghost = Template.bind({})
+Ghost.args = {
+  ghost: true,
+}
